Refactor query components for 63c8ad9e
